CarSaleConsultant/requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-classic
langchain-text-splitters
faiss-cpu
dashscope
gradio